+++ ps_lib/src/main/java/com/edisonwang/ps/lib/ActionRequestEnv.java
@@ -1,5 +1,6 @@
 package com.edisonwang.ps.lib;
 
+import android.content.Context;
 import android.os.Bundle;
 
 /**
@@ -9,10 +10,29 @@ public class ActionRequestEnv implements RequestEnv {
 
     private final Bundle mBundle;
     private final ActionCacheFactory mActionCacheFactory;
+    private final ActionResults mResults;
+    private final EventServiceImpl mServiceImpl;
 
-    public ActionRequestEnv(Bundle bundle, ActionCacheFactory actionCacheFactory) {
+    public ActionRequestEnv(Bundle bundle, ActionCacheFactory actionCacheFactory, EventServiceImpl service) {
         mBundle = bundle;
         mActionCacheFactory = actionCacheFactory;
+        mResults = new ActionResults();
+        mServiceImpl = service;
+    }
+
+    @Override
+    public Bundle getServiceState() {
+        return mServiceImpl.getState();
+    }
+
+    @Override
+    public Context getContext() {
+        return mServiceImpl.getContext();
+    }
+
+    @Override
+    public ActionResults getResults() {
+        return mResults;
     }
 
     @Override
