+++ ps_lib/src/main/java/com/edisonwang/ps/lib/Condition.java
@@ -1,0 +1,16 @@
+package com.edisonwang.ps.lib;
+
+/**
+ *
+ * Conditions that has to isSatisfied after the dependencies are processed.
+ *
+ * @author edi
+ */
+public interface Condition {
+
+    /**
+     * Satisfy the condition, return false if condition is not met and execution should terminate.
+     */
+    boolean isSatisfied(RequestEnv env, ActionRequest request);
+
+}
