+++ ps_lib/src/main/java/com/edisonwang/ps/lib/Requirement.java
@@ -1,0 +1,10 @@
+package com.edisonwang.ps.lib;
+
+import android.os.Parcelable;
+
+/**
+ * @author edi
+ */
+public interface Requirement extends Parcelable {
+    Condition get();
+}
